msm: kgsl: Remove POPP
POPP constantly attempts to lower the GPU's frequency behind the governor's
back in order to save power; however, the GPU governor in use
(msm-adreno-tz) is very good at determining the GPU's load and selecting
an appropriate frequency to run the GPU at.

POPP was created long ago (perhaps when msm-adreno-tz didn't exist or
didn't work so well), so it is clearly deprecated. Remove it.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
Signed-off-by: celtare21 <celtare21@gmail.com>
Signed-off-by: ZyCromerZ <neetroid97@gmail.com>
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>
Signed-off-by: X1LYR4Y <xilyray@gmail.com>