power: qpnp-fg-gen3: Restore the recharge threshold when out of JEITA
tried fixing the fluctuation of charging and termination because
of frequent change of recharge SOC threshold. However, there is
still a corner case existing where the recharge SOC got lowered
after an early charge termination because of JEITA. Charging
won't begin because of the lowered recharge SOC and charge
termination condition doesn't go away. Once the device is out of
JEITA i.e. battery temperature is back to normal, restore the
recharge SOC threshold so that charging can begin quickly.

Change-Id: I97b9228050783822091cbe82b285fa34e9b6d4d0
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>
Signed-off-by: RyuujiX <saputradenny712@gmail.com>
Signed-off-by: ben863 <r.budhiono@gmail.com>
Signed-off-by: Ramadhan Angga Zephyc <ramadhananggayudhanto@gmail.com>
Signed-off-by: a e s x i l y <aesxilymail@gmail.com>